examples/Indic Soundbox AI/requirements.txt
Flask
requests
python-dotenv 